+++ src/bao/lib/contracts.js
