+++ src/contexts/Farms/Farms.tsx
@@ -4,8 +4,8 @@ import { useWallet } from 'use-wallet'
 import useSushi from '../../hooks/useSushi'
 
 import { bnToDec } from '../../utils'
-import { getMasterChefContract, getEarned } from '../../sushi/utils'
-import { getFarms } from '../../sushi/utils'
+import { getMasterChefContract, getEarned } from '../../bao/utils'
+import { getFarms } from '../../bao/utils'
 
 import Context from './context'
 import { Farm } from './types'
