+++ src/hooks/useAllowance.ts
@@ -7,7 +7,7 @@ import { provider } from 'web3-core'
 import { Contract } from 'web3-eth-contract'
 
 import { getAllowance } from '../utils/erc20'
-import { getMasterChefContract } from '../sushi/utils'
+import { getMasterChefContract } from '../bao/utils'
 
 const useAllowance = (lpContract: Contract) => {
   const [allowance, setAllowance] = useState(new BigNumber(0))
