+++ src/hooks/useLockedEarnings.ts
@@ -4,7 +4,7 @@ import { provider } from 'web3-core'
 import BigNumber from 'bignumber.js'
 import { useWallet } from 'use-wallet'
 
-import { getLockedEarned, getSushiContract } from '../sushi/utils'
+import { getLockedEarned, getSushiContract } from '../bao/utils'
 import useSushi from './useSushi'
 import useBlock from './useBlock'
